# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Yantra4D API
  version: 1.0.0
  description: |
    REST API for the Yantra4D parametric 3D design platform.
    Provides OpenSCAD rendering, project management, AI-assisted design,
    Git operations, and analytics.
  license:
    name: AGPL-3.0
    url: https://opensource.org/license/agpl-v3/
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.yantra4d.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Janua-issued JWT token
...
paths:
...
  /api/projects/{slug}/files:
    get:
      tags: [editor]
      summary: List project files
      operationId: listFiles
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: string
    post:
      tags: [editor]
      summary: Create a new file
      operationId: createFile
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filepath:
                  type: string
                content:
                  type: string
              required: [filepath]
      responses:
        "201":
          description: File created
        "400":
          description: Invalid path
        "409":
          description: File already exists
  /api/projects/{slug}/files/{filepath}:
    get:
      tags: [editor]
      summary: Read file content
      operationId: readFile
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: filepath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
        "404":
          description: File not found
    put:
      tags: [editor]
      summary: Write file content
      operationId: writeFile
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: filepath
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        "200":
          description: File written
        "400":
          description: Invalid path
    delete:
      tags: [editor]
      summary: Delete a file
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: filepath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File deleted
        "404":
          description: File not found

  # ── Git ─────────────────────────────────────────────────
  /api/projects/{slug}/git/connect-remote:
    post:
      tags: [git]
      summary: Connect a Git remote
      operationId: connectRemote
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repo_url:
                  type: string
                  format: uri
              required: [repo_url]
      responses:
        "200":
          description: Remote connected
        "400":
          description: Invalid repository URL
  /api/projects/{slug}/git/status:
    get:
      tags: [git]
      summary: Get Git status
      operationId: getGitStatus
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Git status
          content:
            application/json:
              schema:
                type: object
                properties:
                  initialized:
                    type: boolean
                  branch:
                    type: string
                  clean:
                    type: boolean
                  changed_files:
                    type: array
                    items:
                      type: string
                  remote:
                    type: string
  /api/projects/{slug}/git/diff:
    get:
      tags: [git]
      summary: Get Git diff
      operationId: getGitDiff
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Diff output
          content:
            application/json:
              schema:
                type: object
                properties:
                  diff:
                    type: string
  /api/projects/{slug}/git/commit:
    post:
      tags: [git]
      summary: Create a Git commit
      operationId: gitCommit
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
              required: [message]
      responses:
        "200":
          description: Commit created
        "400":
          description: Missing message or nothing to commit
  /api/projects/{slug}/git/push:
    post:
      tags: [git]
      summary: Push to remote
      operationId: gitPush
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pushed successfully
        "400":
          description: No remote configured
  /api/projects/{slug}/git/pull:
    post:
      tags: [git]
      summary: Pull from remote
      operationId: gitPull
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pulled successfully
        "400":
          description: No remote configured

  # ── GitHub ──────────────────────────────────────────────
...